Extract resampling methods
