+++ Project.toml
@@ -3,5 +3,9 @@ uuid = "576499cb-2369-40b2-a588-c64705576edc"
 authors = ["TuringLang"]
 version = "0.1.0"
 
+[deps]
+Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
+
 [compat]
+Distributions = "0.23, 0.24"
 julia = "1.3"
