+++ test/runtests.jl
@@ -2,5 +2,5 @@ using AdvancedPS
 using Test
 
 @testset "AdvancedPS.jl" begin
-    # Write your tests here.
+    @testset "Resampling tests" begin include("resampling.jl") end
 end
